// src/app/common/model/shared-model.ts
export class TreeNode {
  id?: any;
  value?: any;
  label?: any;
  menuPermisFlag?: any;
  permisOrder?: any;
  router?: any;
  parentCode?: any;
  // parentId?: any;
  check?: any;
  color?: any;
  partialSelected?: any;
  remark?: any;
  idt?: any;
  udt?: any;
  type?: any;
  selectable?: any;
  children?: TreeNode[];
}